Use sequoia instead of gpg
This commit updates the depot tool with support for Sequoia PGP. Sequoia
uses gpg-agent as a default backend, thus we can resort to always prefer
Sequoia over Gnupg. For this decision, we check whether the installed
sq binary supports the required cli version.

As a side effect, we renamed the --gpg-user-id argument of the
add-depot-user command to --user-id and added a --fingerprint argument.

genodelabs/goa#104